requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
google-cloud-firestore
google-cloud-storage
requests
resend
